// src/allDtos/claims.dto.ts
export class ClaimsDto {

    id: string;
  
    employeeId: string;
  
    claimDate: string;
  
    claimCategory: string;
  
    claimComment: string;
  
    fileId: string;
  
    fileLink: string;
  
    comment: string;
  
    action: string;
  
    paidBy: string;

    type: string;
  
    status: string;
  
    amount: string;
  
    createdAt: Date;
  
    modifiedAt: Date;
  
    companyId: string;

    requesterId: string;
  }
  